+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+
+- Add `djlint.executablePath` setting for use when `djlint.useVenv` is disabled before the `djlint.pythonPath` fallback.
+
 ## [2025.11.0] - 2025-11-18
 
 - Abort running djLint processes on extension deactivation.